requirements.txt
agno
google-genai
python-dotenv
duckduckgo-search
openai
